Add std.crypto module with 22 primitives; expand NativeFnId to u16
Expand native function ID space from u8 (256 slots) to u16 by changing
NativeFnId's backing type. Fix NativeFnIdArrayLen to use max_enum_value+1
instead of field count — the enum is sparse, so field count < max value,
causing out-of-bounds indexing into native_fn_backing on the WASM
embedding path.

Add std.crypto namespace with 22 functions:
- Hashes: sha256, sha512, blake3, md5, sha1
- MACs: hmac_sha256, hmac_sha512
- AEAD: aes_gcm_seal/open, chacha20poly1305_seal/open,
        xchacha20poly1305_seal/open (24-byte nonce)
- KDF: hkdf_sha256(ikm, salt, info, len), argon2id
- Signatures: ed25519_sign(seed, msg), ed25519_verify(pk, sig, msg)
- DH: x25519(secret, public)
- Passwords: bcrypt_hash, bcrypt_verify
- Utility: rand_bytes, constant_time_equal

AEAD seal output is raw bytes (ciphertext || 16-byte tag); use
std.bytes.len not std.core.len for binary output. Verified against
RFC 5869 §A.1 (HKDF), RFC 7748 §6.1 (X25519), and Zig stdlib test
vectors (Ed25519).

+++ src/lang/module_compile.zig
@@ -88,7 +88,7 @@ pub const HostModuleDesc = struct {
 pub const CapModuleFuncDesc = struct {
     name: []const u8,
     arity: u8,
-    native_id: u8,
+    native_id: u16,
 };
 
 pub const CapModuleDesc = struct {
